Test Suite e primo Integration Test

+++ EnotecaIlGocciolatoioIS/src/main/java/bean/Prodotto.java
@@ -111,7 +111,7 @@ public class Prodotto {
         this.regione = regione;
     }
 
-    public double getGradazione() {
+    public int getGradazione() {
         return gradazione;
     }
